+++ tests/OrderTest.php
@@ -10,6 +10,8 @@ use Lloricode\LaravelPandagoSdk\Facades\LaravelPandagoSdk;
 use function Pest\Faker\faker;
 use function PHPUnit\Framework\assertEquals;
 
+beforeEach(fn() => LaravelPandagoSdk::token()->fake()->fakeGenerate());
+
 it('submit', function () {
     $payloadRequest = <<<FAKE
 {